More changes for CHTC

+++ isce/get_orbits_from_askja.sh
@@ -1,5 +1,6 @@
 #!/bin/bash -e
 ## 2021/06/08 Kurt Feigl
+# 2021/07/20 update to handle ssh keys on CHTC
 
 # have to get the orbits we need from askja.ssec.wisc.edu
 
@@ -10,7 +11,13 @@ rm -f tmp.*
 #ssh -i $HOME/.ssh/id_rsa -l feigl askja.ssec.wisc.edu 'ls /s12/insar/S1/ORBITS' > tmp.fnames0
 #cp -v $HOME/.ssh/id_rsa . 
 #ssh -i ./id_rsa -l feigl askja.ssec.wisc.edu 'ls /s12/insar/S1/ORBITS' > tmp.fnames0
-ssh -i /home/ops/.ssh/id_rsa -l feigl askja.ssec.wisc.edu 'ls /s12/insar/S1/ORBITS' > tmp.fnames0
+if [[ -f /home/ops/.ssh/id_rsa ]]; then
+   ssh -i /home/ops/.ssh/id_rsa -l feigl askja.ssec.wisc.edu 'ls /s12/insar/S1/ORBITS' > tmp.fnames0
+elif [[ -f $HOME/.ssh/id_rsa ]]; then
+   ssh -i $HOME/.ssh/id_rsa -l feigl askja.ssec.wisc.edu 'ls /s12/insar/S1/ORBITS' > tmp.fnames0
+else
+   ssh -l feigl askja.ssec.wisc.edu 'ls /s12/insar/S1/ORBITS' > tmp.fnames0
+fi
 
 
 for slcname in $(ls ../SLC/S1?_IW_SLC*.zip ); do
@@ -32,8 +39,13 @@ for slcname in $(ls ../SLC/S1?_IW_SLC*.zip ); do
 
    for fname1 in ${fnames3}; do
       echo fname1 is $fname1
-      #rsync -av -e "ssh -i $HOME/.ssh/id_rsa -l feigl" askja.ssec.wisc.edu:/s12/insar/S1/ORBITS/${fname1} .
-      rsync -av -e "ssh -i /home/ops/.ssh/id_rsa -l feigl" askja.ssec.wisc.edu:/s12/insar/S1/ORBITS/${fname1} .
+      if [[ -f /home/ops/.ssh/id_rsa ]]; then
+         rsync -av -e "ssh -i /home/ops/.ssh/id_rsa -l feigl" askja.ssec.wisc.edu:/s12/insar/S1/ORBITS/${fname1} .
+      elif [[ -f $HOME/.ssh/id_rsa ]]; then
+         rsync -av -e "ssh -i $HOME/.ssh/id_rsa -l feigl" askja.ssec.wisc.edu:/s12/insar/S1/ORBITS/${fname1} .
+      else
+         rsync -av -e "ssh -l feigl" askja.ssec.wisc.edu:/s12/insar/S1/ORBITS/${fname1} .
+      fi
    done
    #mv -v $fname1 $yyyymmdd  
 done
